feat!: support stubbed properties as well as methods
Changes the method map passed to `stubInterface` and friends to be
a `Partial` of the stubbed type.

This way we can pass values for properties and `Sinon.stub`s for
methods, which is usually better since we freqently want to stub
behaviour as well as return types which we can't do when we're just
passing values.

BREAKING CHANGE: method map values must now be wrapped in `Sinon.stub().returns()`